pyproject.toml
[project]
name = "vibe-to-live-agentic-rag-cohort-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-phoenix-client>=1.21.0",
    "arize-phoenix-evals>=2.5.0",
    "arize-phoenix-otel>=0.13.1",
    "fastapi>=0.119.0",
    "fastembed>=0.7.3",
    "openai>=2.3.0",
    "openai-agents>=0.1.0",
    "openinference-instrumentation>=0.1.40",
    "openinference-instrumentation-openai>=0.1.39",
    "openinference-instrumentation-openai-agents>=1.3.0",
    "openinference-semantic-conventions>=0.1.24",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "pydantic>=2.12.0",
    "qdrant-client>=1.15.1",
    "uvicorn>=0.37.0",
]
